Cleanup C4267 warnings for down-cast for below files:
1. Jshost.cpp
2. PhScriptSite.cpp
3. TestUtilities.cpp
4. WScriptFastDom.cpp
5. HostConfigFlags.cpp
6. HostConfigFlags.h
7. ApiTest.h

+++ private/lib/HostCommon/HostConfigFlags.cpp
@@ -83,7 +83,7 @@ void HostConfigFlags::PrintUsage()
     HostConfigFlags::PrintUsageString();
 }
 
-int HostConfigFlags::FindArg(int argc, _In_reads_(argc) PWSTR argv[], PCWSTR targetArg, int targetArgLen)
+int HostConfigFlags::FindArg(int argc, _In_reads_(argc) PWSTR argv[], PCWSTR targetArg, size_t targetArgLen)
 {
     return FindArg(argc, argv, [=](PCWSTR arg) -> bool
     {
